advanced-api-project/requirements.txt
asgiref==3.9.1
dj-database-url==3.0.1
Django==5.2.5
django-csp==4.0
django-filter==25.1
djangorestframework==3.16.1
packaging==25.0
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
